Added Settings and Resources Component|Added Blog Page Button|Fixed Client Side Exception Error|Added Delete Account Condition for Google Accounts
