+++ website3.0/components/Header.js
@@ -319,7 +319,7 @@ const Header = () => {
             </Link>
           </div>
 
-          {!isCreateBlog || !isProfile && <div className={`flex-grow lg:max-w-[300px] max-lg:max-w-[230px] max-sm:max-w-[220px] max-[445px]:max-w-[100px] max-[400px]:hidden ${theme?"":"text-black"}`}>
+          {!isCreateBlog && !isProfile && <div className={`flex-grow lg:max-w-[300px] max-lg:max-w-[230px] max-sm:max-w-[220px] max-[445px]:max-w-[100px] max-[400px]:hidden ${theme?"":"text-black"}`}>
             <input
             type="text"
             placeholder="Search blogs..."
