+++ website3.0/src/app/api/login/route.js
@@ -28,8 +28,6 @@ export async function POST(req) {
       }
       //comparing the password
    let data1=await bcrypt.compareSync(password,data[0].password)
-   // Checking for incorrect pass 
-   console.log(data1,"ddffffffffffffffffffff")
    if(!data1){
     return NextResponse.json({ success: false,msg:"Incorrect Password"},{status:"200"});
    }
